+++ py_common/time_handling/time_formatters/s_formatter.py
@@ -1,9 +1,9 @@
-from py_common import ABCTimeFormatter
+from .time_formatter import ABCTimeFormatter
 
 
 class SFormatter(ABCTimeFormatter):
     """Formats time based on the SS format."""
-    def format(self, total_seconds: int, round_digits: int) -> str:
+    def format(self, total_seconds: float, round_digits: int) -> str:
         if round_digits is not None:
             seconds = round(total_seconds, round_digits)
         else:
