+++ py_common/time_handling/time_formatters/time_formatter.py
@@ -3,5 +3,5 @@ from abc import ABC, abstractmethod
 
 class ABCTimeFormatter(ABC):
     @abstractmethod
-    def format(self, total_seconds: int, round_digits: int) -> str:
+    def format(self, total_seconds: float, round_digits: int) -> str:
         """Formats time and rounds the seconds to the specified number of digits."""
